remove barrier from list of core-features
The current set of multithreading features together with the
ability of the dynamic interaction with other hosts bring
already everything, which would also be possible by the original
planned barriers. Beside this inplementing the barriers in the
current dynamic base, could make the tree-files much more complicate
and hard to controll in bigger setups.
So I deceided the remove this planned feature.